RelayServer/requirements.txt
bm25>=1.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.9.0
langchain-community>=0.3.5
langchain-openai>=0.2.5
langchain>=0.3.7
openai>=1.53.1
python-dotenv>=1.0.1
pyzmq>=26.2.0
pytz>=2024.2
six>=1.17.0
numpy>=1.26.4
pandas>=2.2.3
pymupdf4llm>=0.0.17